requirements.txt
asgiref==3.11.0
certifi==2025.11.12
cffi==2.0.0
charset-normalizer==3.4.4
cryptography==46.0.3
defusedxml==0.7.1
Django==5.2.9
django-cors-headers==4.9.0
djangorestframework==3.16.1
idna==3.11
mwclient==0.11.0
mwoauth==0.4.0
mwparserfromhell==0.7.2
oauthlib==3.3.1
packaging==25.0
pycparser==2.23
PyJWT==2.10.1
python-social-auth==0.3.6
python3-openid==3.2.0
pywikibot==10.7.3
requests==2.32.5
requests-oauthlib==2.0.0
social-auth-app-django==5.6.0
social-auth-core==4.8.1
sqlparse==0.5.4
urllib3==2.6.1
